Veneer-migrate 4 repair-exhausted/invalid dispatch methods (batch 18)
Phase 2 batch 18 for Issue #682. Migrates 4 more impl Agent dispatch
methods including the repair-exhausted emit pair and the two main
repair-pass driver methods.

Moved (turn.rs impl Agent -> verifier_orchestration.rs free fn):

  - emit_safe_stop_report_for_repair_exhausted (3 LOC) — thin shell
    that delegates to emit_repair_safe_stop_report with the
    RepairExhausted stop reason.
  - maybe_emit_repair_exhausted_from_promotion (8 LOC) — Codex
    CB-002 chokepoint that observes PromotionResult.all_clusters_exhausted
    and conditionally emits the repair_exhausted safe-stop report.
  - run_verifier_repair_pass_and_apply (93 LOC) — top-level
    verifier-repair pass driver: prepares the pass, iterates up to
    VERIFIER_REPAIR_PASS_ATTEMPT_LIMIT attempts with wall-clock
    timeout enforcement, requests the patch provider LLM, dispatches
    attempt handling, and emits the final Invalid event with the
    bounded last_invalid_outcome.
  - record_controller_verifier_repair_invalid (83 LOC) — records a
    rejected repair-intent attempt: sanitizes the error, applies
    PatchRejected lifecycle events, records the attempt outcome with
    PromotionResult observation, and chains
    maybe_emit_repair_exhausted_from_promotion.

Adjustments:

  * verifier_orchestration.rs:
    - Veneers call `agent.emit_repair_safe_stop_report(...)` and
      `agent.prepare_verifier_repair_pass(...)` /
      `agent.handle_verifier_repair_pass_attempt(...)` /
      `agent.verifier_repair_pass_wall_clock_timeout_error(...)`
      directly via the &mut Agent first arg.
    - All deps (RepairAttemptOutcome, PromotionResult,
      RejectedAttemptReason, sanitize_repair_job_text_with_char_cap,
      rejected_reason_for_repair_*, lifecycle_event_for_repair_error,
      malformed_repair_attempt_outcome_for_active_target,
      verifier_repair_effective_target_hint, RepairAttemptKey,
      RepairJobEvent, compact_verifier_failure_text,
      VERIFIER_REPAIR_PASS_ATTEMPT_LIMIT,
      verifier_repair_pass_attempt_timeout_secs,
      verifier_repair_pass_retry_message) reachable via
      super::repair_job, super::repair_driver, or already imported
      from earlier batches.

  * turn.rs:
    - Promoted 4 Agent methods to pub(super) so the relocated
      veneers can call them via agent.method(...):
      prepare_verifier_repair_pass, handle_verifier_repair_pass_attempt,
      verifier_repair_pass_wall_clock_timeout_error,
      emit_repair_safe_stop_report.
    - Deleted the 4 original impl Agent method bodies (187 LOC).
    - Rewrote 5 call sites (4 fns x 1+ callers) to
      super::verifier_orchestration::*(self, ...) inside turn.rs,
      including the 2 actor_loop_flow.rs call sites that go through
      `agent.emit_safe_stop_report_for_repair_exhausted()` and
      `agent.maybe_emit_repair_exhausted_from_promotion(...)`.
    - Updated 2 callers in loop_run.rs (test seam helpers) to
      `verifier_orchestration::*(agent, ...)`.
    - Dropped now-unused turn.rs imports
      VERIFIER_REPAIR_PASS_ATTEMPT_LIMIT,
      verifier_repair_pass_attempt_timeout_secs,
      verifier_repair_pass_retry_message,
      malformed_repair_attempt_outcome_for_active_target,
      verifier_repair_effective_target_hint, and
      compact_verifier_failure_text.
    - Re-added verifier_repair_pass_retry_message +
      verifier_repair_effective_target_hint inside mod tests'
      explicit `use crate::agent::loop_run::*` block.
    - Re-added verifier_repair_pass_attempt_timeout_secs as a
      `use super::super::repair_driver::*` line inside
      mod repair_lifecycle_event_tests.
    - Rewrote `super::malformed_repair_attempt_outcome_for_active_target(...)`
      test fixture call to use the full
      `super::super::repair_job::malformed_repair_attempt_outcome_for_active_target(...)`
      path.
    - Cleaned an orphaned doc-comment block above the now-deleted
      `maybe_emit_repair_exhausted_from_promotion` method (clippy
      empty_line_after_doc_comments under -D warnings).
    - Updated 3 include_str! test markers in
      v0421_repair_runner_contract_tests to read
      run_verifier_repair_pass_and_apply's body from
      verifier_orchestration.rs (start marker
      `\npub(super) fn run_verifier_repair_pass_and_apply(`,
      end marker
      `\npub(super) fn record_controller_verifier_repair_invalid(`).
      Relaxed the body.contains("&super::task_contract::RecoveryTargetHint")
      assertion to body.contains("&RecoveryTargetHint") since the
      type is now imported unqualified in verifier_orchestration.rs.

Metrics: turn.rs 32,295 -> 32,090 (-205);
verifier_orchestration.rs 2,809 -> 2,965 (+156). Cumulative Issue
#682 progress: turn.rs 34,958 -> 32,090 (-2,868, ~96% of Phase 2
target). Acceptance gap: 32,090 -> 32,000 = ~90 LOC remaining.
